requirements.txt
django
pymongo
Werkzeug
PyJWT